requirements/base.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.2.3
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-jazzmin==3.0.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
drf-standardized-errors==0.15.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
psycopg==3.2.4
psycopg2-binary==2.9.10
pytz==2025.2
pandas==2.3.1
numpy
Faker==37.5.3
